+++ examples/miniapps/movie_lister/app_csv.py
@@ -26,17 +26,17 @@ import dependency_injector.providers as providers
 class MyMoviesModule(containers.DeclarativeContainer):
     """IoC container for overriding movies module component providers."""
 
-    movie_finder = providers.Factory(movies.finders.CsvMovieFinder,
-                                     csv_file_path=settings.MOVIES_CSV_PATH,
-                                     delimiter=',',
-                                     **movies.MoviesModule.movie_finder.kwargs)
+    finder = providers.Factory(movies.finders.CsvMovieFinder,
+                               csv_file_path=settings.MOVIES_CSV_PATH,
+                               delimiter=',',
+                               **movies.MoviesModule.finder.kwargs)
 
 
 class CsvApplication(containers.DeclarativeContainer):
     """IoC container of csv application component providers."""
 
     main = providers.Callable(example.main.main,
-                              movie_lister=movies.MoviesModule.movie_lister)
+                              movie_lister=movies.MoviesModule.lister)
 
     init_db = providers.Callable(example.db.init_csv,
                                  movies_data=fixtures.MOVIES_SAMPLE_DATA,